auth-server/requirements.txt
pycryptodome
requests
flask
